Facebook-clone/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-heroku==0.3.1
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
dnspython==2.1.0
email-validator==1.1.2
Faker==8.2.0
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
sentry-sdk==1.1.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.4
whitenoise==5.2.0
